requirements.txt
django
flake8
flake8-quotes
flake8-variables-names
pep8-naming
django-debug-toolbar
djangorestframework
djangorestframework-simplejwt
drf-spectacular
Pillow
psycopg==3.1.12
psycopg-binary==3.1.12
python-slugify